+++ data/db/BvXs9Tw4ZsURV6knGwhKa/alias.json.js
@@ -1,2 +1,0 @@
-jsonjs.data['alias'] = 
-[["id","table","alias"],[1,"institution","owner"],[2,"institution","manager"]]