+++ data/db/BvXs9Tw4ZsURV6knGwhKa/filter.json.js
@@ -1,2 +1,0 @@
-jsonjs.data['filter'] = 
-[["id","name"],["open_data","Open Data"],["closed_data","Closed Data"]]