+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - changed : get last modification timestamp directly from jsonjsdb instead of info file
+- changed : get filter and alias info from config file instead of filter and alias files
 - fixed : typo in about_meta
 - fixed : doc preview a pixel to height
 - fixed : tabs_body overflow hidden